* commit '6399980a1d22ecf2b444c106e02e36234a9f8263': (172 commits)
  temp
  Modified - changed IDs to fix tests
  Modified - planning-exp-book.html: added hidden input to set user in frontend - planningtool.html: added column to show user name of experiment
  bugfix
  device fix
  Tests failures with / fixed
  Added Controller Tests for experiment types
  Fixed homecontroller coverage
  Stock Statistics show correct current stock
  Modified - StepController: fixed nullpointer exception when saving experiment without mixture
  Modified - HomeControllerTests: Changed ID form 32 to 33 and added tests on size of sudentSteps, userSteps, userExperiment and studentExperiment.
  Modified - StepController: fixed check of office hours at 17:00, modified methods because of problem with testing inside StepControllerTests
  uncommented @profile in databaseloader
  pulled sprint2 before pull request
  Tests 100% coverage
  Added Tests for StepType repo,service and experiment Type service
  When user books experiment, materials are withdrawn from stock. If insufficient stock, experiment is not booked, user returns to planning page, notification shows up.
  Steptype with name problem fixed
  Modified - planning-exp-book.html: Mixture amount can be only number
  Modified - DatabaseLoader: modified saving of Experiment
  ...

+++ src/main/java/be/uantwerpen/labplanner/Controller/RoleController.java
@@ -1,19 +1,33 @@
 package be.uantwerpen.labplanner.Controller;
 
 
+import be.uantwerpen.labplanner.Service.OwnPrivilegeService;
+import be.uantwerpen.labplanner.common.model.users.Privilege;
 import be.uantwerpen.labplanner.common.model.users.Role;
+import be.uantwerpen.labplanner.common.model.users.User;
 import be.uantwerpen.labplanner.common.service.users.PrivilegeService;
 import be.uantwerpen.labplanner.common.service.users.RoleService;
+import be.uantwerpen.labplanner.common.service.users.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.servlet.LocaleResolver;
+import org.springframework.web.servlet.i18n.SessionLocaleResolver;
 
 import javax.validation.Valid;
 import java.beans.MethodDescriptor;
+import java.net.ContentHandler;
+import java.util.List;
+import java.util.Locale;
+import java.util.ResourceBundle;
 
 @Controller
 public class RoleController {
@@ -22,50 +36,100 @@ public class RoleController {
     private RoleService roleService;
 
     @Autowired
-    private PrivilegeService privilegeService;
+    private OwnPrivilegeService privilegeService;
 
+    @Autowired
+    private UserService userService;
+
+
+    //Populate
+    @ModelAttribute("allRoles")
+    public Iterable<Role> populateRoles(){return roleService.findAll();}
+
+    @ModelAttribute("allPrivileges")
+    public Iterable<Privilege> populatePrivileges(){return privilegeService.findAll();}
+
+    @PreAuthorize("hasAnyAuthority('User Management')")
     @RequestMapping(value = "/usermanagement/roles",method = RequestMethod.GET)
     public String showRoles(final ModelMap model){
         model.addAttribute("allRoles",roleService.findAll());
-        return "/Roles/role-list";
+        return "Roles/role-list";
     }
 
+    @PreAuthorize("hasAnyAuthority('User Management')")
     @RequestMapping(value = "/usermanagement/roles/put", method = RequestMethod.GET)
     public String ViewCreateRole(final ModelMap model){
         model.addAttribute("allPrivileges",privilegeService.findAll());
         model.addAttribute(new Role(""));
-        return "/Roles/role-manage";
+        return "Roles/role-manage";
     }
 
+    @PreAuthorize("hasAnyAuthority('User Management')")
     @RequestMapping(value = "/usermanagement/roles/{id}",method = RequestMethod.GET)
     public String viewEditRole(@PathVariable long id, final ModelMap model){
         model.addAttribute("allPrivileges",privilegeService.findAll());
         model.addAttribute("role",roleService.findById(id));
-        return "/Roles/role-manage";
+        return "Roles/role-manage";
     }
 
+    @PreAuthorize("hasAnyAuthority('User Management')")
     @RequestMapping(value = {"/usermanagement/roles/","/usermanagement/roles/{id}"},method = RequestMethod.POST)
     public String addRole(@Valid Role role, BindingResult result, final ModelMap model) {
-        if (result.hasErrors()) {
+        if ((result.hasErrors()) || (role.getName() == null)||( role.getName().trim().equals(""))) {
             model.addAttribute("allPrivileges", privilegeService.findAll());
-            return "/Roles/role-manage";
+            model.addAttribute("roleInUse", ResourceBundle.getBundle("messages", LocaleContextHolder.getLocale()).getString("roles.errorAdd"));
+            return "Roles/role-manage";
         }
         if (role.getId() == null) {
             if (roleService.findByName(role.getName()).isPresent()) {
                 model.addAttribute("allPrivileges", privilegeService.findAll());
-                model.addAttribute("roleInUse", "Role " + role.getName() + " is already in use!");
-                return "/Roles/role-manage";
+                model.addAttribute("roleInUse", ResourceBundle.getBundle("messages",LocaleContextHolder.getLocale()).getString("roles.errorUnique"));
+                return "Roles/role-manage";
             }
+            //trim input and save
+            role.setName(role.getName().trim());
             roleService.save(role);
             return "redirect:/usermanagement/roles";
         }
 
+        Role tempRole = roleService.findById(role.getId()).orElse(null);
+        if (!tempRole.getName().equals(role.getName())){
+            if (roleService.findByName(role.getName()).isPresent()){
+                model.addAttribute("allPrivileges", privilegeService.findAll());
+                model.addAttribute("roleInUse", ResourceBundle.getBundle("messages",LocaleContextHolder.getLocale()).getString("roles.errorUnique"));
+                return "Roles/role-manage";
+            }
+            //trim input and save
+            role.setName(role.getName().trim());
+            roleService.save(role);
+            return "redirect:/usermanagement/roles";
+        }
+        //trim input and save
+        role.setName(role.getName().trim());
         roleService.save(role);
         return "redirect:/usermanagement/roles";
     }
 
+    @PreAuthorize("hasAnyAuthority('User Management')")
     @RequestMapping(value = "/usermanagement/roles/{id}/delete",method = RequestMethod.GET)
     public String deleteRole(@PathVariable long id, final ModelMap model){
+        //print messages based in current locale
+        List<User> allUsers = userService.findAll();
+        boolean isUsed = false;
+        for (User user : allUsers){
+            for (Role role : user.getRoles()){
+                if (role.getId()==id){
+                    isUsed = true;
+                }
+            }
+        }
+        if (isUsed){
+            model.addAttribute("allRoles",roleService.findAll());
+
+            //get message from resourceBundle
+            model.addAttribute("inUseError",ResourceBundle.getBundle("messages",LocaleContextHolder.getLocale()).getString("roles.errorDelete"));
+            return "Roles/role-list";
+        }
         roleService.deleteById(id);
         model.clear();
         return "redirect:/usermanagement/roles";
